+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 - Node.js: Support `args.notApplicable(reason)`
   (https://github.com/pulumi/pulumi-policy/pull/409).
 
+- Python: Add support for new metadata fields for policy packs and policies
+  (https://github.com/pulumi/pulumi-policy/pull/410).
+
 ---
 
 ## 1.17.0 (2025-07-29)